refactor: extract books and borrowings from library into separate apps
