+++ borrowings/serializers.py
@@ -1,34 +1,9 @@
 from django.utils import timezone
 from rest_framework import serializers
 
-from library.models import Book, Borrowing
+from books.serializers import BookSerializer
+from borrowings.models import Borrowing
 from user.serializers import UserSerializer
-
-
-class BookSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Book
-        fields = ("id", "title", "author", "cover", "inventory", "daily_fee")
-
-    @staticmethod
-    def validate_inventory(value):
-        if value < 0:
-            raise serializers.ValidationError("Inventory cannot be negative.")
-        return value
-
-    @staticmethod
-    def validate_daily_fee(value):
-        if value < 0:
-            raise serializers.ValidationError("Daily fee cannot be negative.")
-        return value
-
-
-class BookListSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Book
-        fields = ("id", "title", "author", "daily_fee")
 
 
 class BorrowingSerializer(serializers.ModelSerializer):
