+++ config/urls.py
@@ -21,6 +21,10 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/user/", include("user.urls", namespace="user")),
-    path("api/library/", include("library.urls", namespace="library")),
+    path("api/books/", include("books.urls", namespace="books")),
+    path(
+        "api/borrowings/",
+        include("borrowings.urls", namespace="borrowings"),
+    ),
     path("__debug__/", include("debug_toolbar.urls")),
 ]
